move sanford config and hosts util classes into their own files
This is just a reorg of this logic that formalizes it a bit.  This
is prep for adding new features to the config class.  This is also
gets these classes more up to our convention of formalizing logic
like this in its own file.
